# refactor: use Arc<K> Arc<V> for in-memory cache handle