--- a/foyer-common/src/arc_key_hash_map.rs
+++ b/foyer-common/src/arc_key_hash_map.rs
@@ -25,13 +25,26 @@ use hashbrown::hash_table::{HashTable, IntoIter as HashTableIntoIter};
 
 pub use hashbrown::hash_table::Entry;
 
-pub struct HashTableEntry<K, V> {
+use crate::arcable::Arcable;
+
+pub struct ArcKeyHashMapEntry<K, V> {
     key: Arc<K>,
     value: V,
     hash: u64,
 }
 
-impl<K, V> HashTableEntry<K, V> {
+impl<K, V> ArcKeyHashMapEntry<K, V> {
+    pub fn new<AK>(hash: u64, key: AK, value: V) -> Self
+    where
+        AK: Into<Arcable<K>>,
+    {
+        Self {
+            key: key.into().into_arc(),
+            value,
+            hash,
+        }
+    }
+
     pub fn key(&self) -> &Arc<K> {
         &self.key
     }
@@ -40,17 +53,158 @@ impl<K, V> HashTableEntry<K, V> {
         &self.value
     }
 
+    pub fn key_mut(&mut self) -> &mut Arc<K> {
+        &mut self.key
+    }
+
+    pub fn value_mut(&mut self) -> &mut V {
+        &mut self.value
+    }
+
     pub fn take(self) -> (Arc<K>, V) {
         (self.key, self.value)
     }
 }
 
+pub struct RawArcKeyHashMap<K, V> {
+    inner: HashTable<ArcKeyHashMapEntry<K, V>>,
+}
+
+impl<K, V> Debug for RawArcKeyHashMap<K, V> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("RawArcHashMap").finish()
+    }
+}
+
+impl<K, V> Default for RawArcKeyHashMap<K, V> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl<K, V> RawArcKeyHashMap<K, V> {
+    pub fn new() -> Self {
+        Self {
+            inner: HashTable::new(),
+        }
+    }
+}
+
+impl<K, V> Deref for RawArcKeyHashMap<K, V> {
+    type Target = HashTable<ArcKeyHashMapEntry<K, V>>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.inner
+    }
+}
+
+impl<K, V> DerefMut for RawArcKeyHashMap<K, V> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.inner
+    }
+}
+
+impl<K, V> RawArcKeyHashMap<K, V>
+where
+    K: Hash + Eq,
+{
+    pub fn insert_with_hash<AK>(&mut self, hash: u64, key: AK, value: V) -> Option<V>
+    where
+        AK: Into<Arcable<K>>,
+    {
+        let key = key.into().into_arc();
+
+        match self.inner.entry(
+            hash,
+            |entry| entry.key.as_ref().borrow() == key.as_ref(),
+            |entry| entry.hash,
+        ) {
+            Entry::Occupied(mut o) => {
+                let mut e = ArcKeyHashMapEntry { key, value, hash };
+                std::mem::swap(o.get_mut(), &mut e);
+                Some(e.value)
+            }
+            Entry::Vacant(v) => {
+                v.insert(ArcKeyHashMapEntry { key, value, hash });
+                None
+            }
+        }
+    }
+
+    pub fn get_with_hash<Q>(&self, hash: u64, key: &Q) -> Option<&V>
+    where
+        K: Borrow<Q>,
+        Q: Hash + Eq + ?Sized,
+    {
+        self.inner
+            .find(hash, |entry| entry.key.as_ref().borrow() == key)
+            .map(|entry| &entry.value)
+    }
+
+    pub fn remove_with_hash<Q>(&mut self, hash: u64, key: &Q) -> Option<V>
+    where
+        K: Borrow<Q>,
+        Q: Hash + Eq + ?Sized,
+    {
+        match self
+            .inner
+            .entry(hash, |entry| entry.key.as_ref().borrow() == key, |entry| entry.hash)
+        {
+            Entry::Occupied(o) => {
+                let (entry, _v) = o.remove();
+                Some(entry.value)
+            }
+            Entry::Vacant(_) => None,
+        }
+    }
+
+    pub fn entry_with_hash<AK>(&mut self, hash: u64, key: AK) -> Entry<'_, ArcKeyHashMapEntry<K, V>>
+    where
+        AK: Into<Arcable<K>>,
+    {
+        let key = key.into().into_arc();
+
+        self.inner.entry(
+            hash,
+            |entry| entry.key.as_ref().borrow() == key.as_ref(),
+            |entry| entry.hash,
+        )
+    }
+
+    pub fn drain(&mut self) -> impl Iterator<Item = (Arc<K>, V)> + '_ {
+        self.inner.drain().map(|entry| (entry.key, entry.value))
+    }
+}
+
+pub struct RawIntoIter<K, V> {
+    inner: HashTableIntoIter<ArcKeyHashMapEntry<K, V>>,
+}
+
+impl<K, V> Iterator for RawIntoIter<K, V> {
+    type Item = (Arc<K>, V);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.inner.next().map(|entry| (entry.key, entry.value))
+    }
+}
+
+impl<K, V> IntoIterator for RawArcKeyHashMap<K, V> {
+    type Item = (Arc<K>, V);
+    type IntoIter = RawIntoIter<K, V>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        RawIntoIter {
+            inner: self.inner.into_iter(),
+        }
+    }
+}
+
 pub struct ArcKeyHashMap<K, V, S = RandomState> {
-    inner: HashTable<HashTableEntry<K, V>>,
+    raw: RawArcKeyHashMap<K, V>,
     build_hasher: S,
 }
 
-impl<K, V> Debug for ArcKeyHashMap<K, V, RandomState> {
+impl<K, V, S> Debug for ArcKeyHashMap<K, V, S> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("ArcHashMap").finish()
     }
@@ -71,23 +225,23 @@ impl<K, V> ArcKeyHashMap<K, V, RandomState> {
 impl<K, V, S> ArcKeyHashMap<K, V, S> {
     pub fn new_with_hasher(build_hasher: S) -> Self {
         Self {
-            inner: HashTable::new(),
+            raw: RawArcKeyHashMap::new(),
             build_hasher,
         }
     }
 }
 
 impl<K, V, S> Deref for ArcKeyHashMap<K, V, S> {
-    type Target = HashTable<HashTableEntry<K, V>>;
+    type Target = HashTable<ArcKeyHashMapEntry<K, V>>;
 
     fn deref(&self) -> &Self::Target {
-        &self.inner
+        &self.raw
     }
 }
 
 impl<K, V, S> DerefMut for ArcKeyHashMap<K, V, S> {
     fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.inner
+        &mut self.raw
     }
 }
 
@@ -96,23 +250,13 @@ where
     K: Hash + Eq,
     S: BuildHasher + Send + Sync + 'static,
 {
-    pub fn insert(&mut self, key: Arc<K>, value: V) -> Option<V> {
-        let hash = self.build_hasher.hash_one(&key);
-        match self.inner.entry(
-            hash,
-            |entry| entry.key.as_ref().borrow() == key.as_ref(),
-            |entry| entry.hash,
-        ) {
-            Entry::Occupied(mut o) => {
-                let mut e = HashTableEntry { key, value, hash };
-                std::mem::swap(o.get_mut(), &mut e);
-                Some(e.value)
-            }
-            Entry::Vacant(v) => {
-                v.insert(HashTableEntry { key, value, hash });
-                None
-            }
-        }
+    pub fn insert<AK>(&mut self, key: AK, value: V) -> Option<V>
+    where
+        AK: Into<Arcable<K>>,
+    {
+        let key = key.into().into_arc();
+        let hash = self.build_hasher.hash_one(key.as_ref());
+        self.raw.insert_with_hash(hash, key, value)
     }
 
     pub fn get<Q>(&self, key: &Q) -> Option<&V>
@@ -121,9 +265,7 @@ where
         Q: Hash + Eq + ?Sized,
     {
         let hash = self.build_hasher.hash_one(key);
-        self.inner
-            .find(hash, |entry| entry.key.as_ref().borrow() == key)
-            .map(|entry| &entry.value)
+        self.raw.get_with_hash(hash, key)
     }
 
     pub fn remove<Q>(&mut self, key: &Q) -> Option<V>
@@ -132,37 +274,32 @@ where
         Q: Hash + Eq + ?Sized,
     {
         let hash = self.build_hasher.hash_one(key);
-        match self
-            .inner
-            .entry(hash, |entry| entry.key.as_ref().borrow() == key, |entry| entry.hash)
-        {
-            Entry::Occupied(o) => {
-                let (entry, _v) = o.remove();
-                Some(entry.value)
-            }
-            Entry::Vacant(_) => None,
-        }
+        self.raw.remove_with_hash(hash, key)
     }
 
-    pub fn entry(&mut self, key: Arc<K>) -> Entry<'_, HashTableEntry<K, V>> {
-        let hash = self.build_hasher.hash_one(&key);
-        self.inner.entry(
-            hash,
-            |entry| entry.key.as_ref().borrow() == key.as_ref(),
-            |entry| entry.hash,
-        )
+    pub fn entry<AK>(&mut self, key: AK) -> Entry<'_, ArcKeyHashMapEntry<K, V>>
+    where
+        AK: Into<Arcable<K>>,
+    {
+        let key = key.into().into_arc();
+        let hash = self.build_hasher.hash_one(key.as_ref());
+        self.raw.entry_with_hash(hash, key)
+    }
+
+    pub fn drain(&mut self) -> impl Iterator<Item = (Arc<K>, V)> + '_ {
+        self.raw.drain()
     }
 }
 
 pub struct IntoIter<K, V> {
-    inner: HashTableIntoIter<HashTableEntry<K, V>>,
+    inner: RawIntoIter<K, V>,
 }
 
 impl<K, V> Iterator for IntoIter<K, V> {
     type Item = (Arc<K>, V);
 
     fn next(&mut self) -> Option<Self::Item> {
-        self.inner.next().map(|entry| (entry.key, entry.value))
+        self.inner.next()
     }
 }
 
@@ -172,7 +309,7 @@ impl<K, V, S> IntoIterator for ArcKeyHashMap<K, V, S> {
 
     fn into_iter(self) -> Self::IntoIter {
         IntoIter {
-            inner: self.inner.into_iter(),
+            inner: self.raw.into_iter(),
         }
     }
 }