@MrCroxx 2024-04-22
## Benchmark

### cmd

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-storage-bench --dir /p44pro/foyer --capacity 204800 --file-size 64 --get-range 10000 --flushers 12 --reclaimers 12 --time 180 --writers 64 --w-rate 8 --ticket-insert-rate-limit 500 --readers 1024 --r-rate 1 --runtime --metrics --compression none --distribution zipf --distribution-zipf-s 0.8
```

### main (as https://github.com/MrCroxx/foyer/pull/371#issuecomment-2066754394)

```
Total:
disk total iops: 4247.5
disk total throughput: 511.8 MiB/s
disk read iops: 783.1
disk read throughput: 52.0 MiB/s
disk write iops: 3464.3
disk write throughput: 459.8 MiB/s
insert iops: 6922.4/s
insert throughput: 432.7 MiB/s
insert lat p50: 36us
insert lat p90: 76us
insert lat p99: 327us
insert lat p999: 847us
insert lat p9999: 1479us
insert lat p99999: 2847us
insert lat pmax: 12095us
get iops: 19723.8/s
get miss: 16.95%
get throughput: 1.0 GiB/s
get hit lat p50: 931us
get hit lat p90: 1711us
get hit lat p99: 2623us
get hit lat p999: 6687us
get hit lat p9999: 11903us
get hit lat p99999: 20479us
get hit lat pmax: 24703us
get miss lat p50: 579us
get miss lat p90: 927us
get miss lat p99: 1375us
get miss lat p999: 4287us
get miss lat p9999: 24959us
get miss lat p99999: 25471us
get miss lat pmax: 25983us
```

### PR

```
Total:
disk total iops: 7969.2
disk total throughput: 527.7 MiB/s
disk read iops: 1006.3
disk read throughput: 66.8 MiB/s
disk write iops: 6963.0
disk write throughput: 460.9 MiB/s
insert iops: 6879.6/s
insert throughput: 430.0 MiB/s
insert lat p50: 38us
insert lat p90: 73us
insert lat p99: 385us
insert lat p999: 871us
insert lat p9999: 1711us
insert lat p99999: 2655us
insert lat pmax: 3199us
get iops: 16335.5/s
get miss: 0.82%
get throughput: 1.0 GiB/s
get hit lat p50: 891us
get hit lat p90: 1991us
get hit lat p99: 2975us
get hit lat p999: 6271us
get hit lat p9999: 8511us
get hit lat p99999: 9087us
get hit lat pmax: 9343us
get miss lat p50: 125us
get miss lat p90: 347us
get miss lat p99: 767us
get miss lat p999: 1431us
get miss lat p9999: 2335us
get miss lat p99999: 2463us
get miss lat pmax: 2463us
```